server/requirements.txt
Flask_Cors==3.0.9
stripe==2.74.0
Flask==2.0.3
passlib==1.7.4
Flask_JWT_Extended==3.24.1
Werkzeug==2.0.0
pymongo==3.10.1
python-dotenv==0.20.0
PyJWT==1.7.1
Flask-Mail==0.9.1
pyaml==20.4.0
